// generated/src/app/api-models/commission-matrix-header.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CommissionMatrixHeader {
  matrixDef: string;
  matrixDescription: string;
  matrixCalcMethod: string;
  matrixDeterminant: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}